refactor: modify project structure, add loading application version from pom.xml
