+++ src/main/java/org/example/learningsystem/course/dto/CourseRequestDto.java
@@ -9,7 +9,7 @@ import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
-import static org.example.learningsystem.util.format.DataFormatUtils.DATE_TIME_FORMAT;
+import static org.example.learningsystem.core.util.format.DataFormatUtils.DATE_TIME_FORMAT;
 
 public record CourseRequestDto(
         UUID id,
