+++ src/main/java/org/example/learningsystem/student/dto/StudentRequestDto.java
@@ -8,7 +8,7 @@ import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.UUID;
 
-import static org.example.learningsystem.util.format.DataFormatUtils.DATE_FORMAT;
+import static org.example.learningsystem.core.util.format.DataFormatUtils.DATE_FORMAT;
 
 public record StudentRequestDto(
         UUID id,
