+++ src/main/java/org/example/learningsystem/student/exception/InsufficientBirthDateException.java
@@ -1,7 +1,7 @@
 package org.example.learningsystem.student.exception;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
-import org.example.learningsystem.util.format.DataFormatUtils;
+import org.example.learningsystem.core.util.format.DataFormatUtils;
 
 import java.time.LocalDate;
